+++ app/src/main/java/edu/three/objects/RaycastItem.java
@@ -7,13 +7,13 @@ import edu.three.math.Vector2;
 import edu.three.math.Vector3;
 
 public class RaycastItem {
-    public float distance;
+    public double distance;
     public Vector3 point;
     public Object3D object;
     public int faceIndex;
     public Face3 face;
 
-    public float distanceToRay;
+    public double distanceToRay;
     public int index;
 
     public Vector2 uv;
